Fixed the next build errors with dmd v2.067.0: cmdopt.d(96): Error: undefined identifier 'c' lexer.d(224): Error: function lexer.Lexer.getForward (ulong position, int num) is not callable using argument types (ulong, ulong) session.d(66): Error: undefined identifier 'c'

+++ lexer.d
@@ -227,7 +227,7 @@ final class Lexer
         return "";
     }
 
-    string getForward(size_t position, int num)
+    string getForward(size_t position, ulong num)
     {
         if (position + num < source.length)
              return source[position..position+num];
